% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/liftover.r
\name{liftover}
\alias{liftover}
\title{Remap genomic coordinates to a different genome assembly}
\usage{
liftover(x, from = genome(x), to, chain.path = getOption("chain.path"))
}
\arguments{
\item{x}{GRanges object}

\item{from}{genome assembly to which x is currently mapped}

\item{to}{genome assembly to which x should be mapped}
}
\description{
Remap genomic coordinates to a different genome assembly
}

