\name{liftoverr-package}
\alias{liftoverr-package}
\alias{liftoverr}
\docType{package}
\title{
\packageTitle{liftoverr}
}
\description{
\packageDescription{liftoverr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{liftoverr}
\packageIndices{liftoverr}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{liftoverr}

Maintainer: \packageMaintainer{liftoverr}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
