\name{liftover}
\alias{liftover}
\title{Remap genomic coordinates to a different genome assembly}
\usage{
liftover(x, from = genome(x), to)
}
\arguments{
  \item{x}{GRanges object}

  \item{from}{genome assembly to which x is currently
  mapped}

  \item{to}{genome assembly to which x should be mapped}
}
\description{
Remap genomic coordinates to a different genome assembly
}

