% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftoveR.R
\name{liftover}
\alias{liftover}
\title{Lift over coordinates of variants from one genome build to another, by realignment}
\usage{
liftover(chrom, start, end = NULL, originalBuild, newBuild, varnames = NULL,
  lengthSides = 25, tmpdir = NULL)
}
\arguments{
\item{chrom}{For each variant, the chromosome on which it resides}

\item{start, end}{Physical coordinates of variants (in 1-based indexing. see \url{http://genomewiki.ucsc.edu/index.php/Coordinate_Transforms})}

\item{originalBuild}{Path to the fasta sequence of the build from which the variants originate}

\item{newBuild}{Path to the fasta sequence of the build to which the variants will be lifted over}

\item{varnames}{Variant names. If missing, variants will be given names based as chromosome_position in original build}

\item{lengthSides}{number of bases which will be extracted from each side of the variant}
}
\value{
A data frame containing the variants with their positions in the new build, based on the following scheme:
\describe{
\item{name}{Name of the variant}
\item{chrom_orig, start_orig, end_orig}{The original position supplied to function.}
\item{chrom, start, end}{Coordinates of alignment in the new build}
}
\strong{Note:} Currently, a log file is saved to the current working directory. This is a feature of the QuasR package used for mapping, and it is currently not possible to disable it.
}
\description{
Lift over coordinates of variants from one genome build to another, by realignment
}

