% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftoveR.R
\name{liftover}
\alias{liftover}
\title{Lift over coordinates of variants from one genome build to another, by realignment}
\usage{
liftover(chrom, start, end = NULL, originalBuild, newBuild, varnames = NULL,
  lengthSides = 25, maxMismatches = 0, tmpdir = NULL,
  aligner = "Rsubread", memlimit = 8000, maxAllowedLength = 500,
  blastArgs = "-max_hsps 1 -parse_deflines -outfmt 15 -max_target_seqs 1")
}
\arguments{
\item{chrom}{For each variant, the chromosome on which it resides}

\item{start, end}{Physical coordinates of variants (in 1-based indexing. see \url{http://genomewiki.ucsc.edu/index.php/Coordinate_Transforms})}

\item{originalBuild}{Path to the fasta sequence of the build from which the variants originate}

\item{newBuild}{Path to the fasta sequence of the build to which the variants will be lifted over}

\item{varnames}{Variant names. If missing, variants will be given names based as chromosome_position in original build}

\item{lengthSides}{Number of bases which will be extracted from each side of the variant}

\item{maxMismatches}{Corresponds to the number of mismatches. See the vignette for more information on how this is specified for each aligner. {Not implemented for blastn}}

\item{tmpdir}{Working directory for sequence alignment etc. Defaults to a temporary directory created automatically by the current R session.}

\item{aligner}{use Rbowtie, Rsubread or blastn for alignment? Recommended that be set to blastn, but could be slow on large genomes, and unneccessary for aligning simple polymorphisms.}

\item{memlimit}{Memory (in MB) limit for building the Rsubread index (relevant only if Rsubread is TRUE). Defaults to 8000MB, which should allow mammalian genomes to be built with high efficiency for fast alignment (see Rsubread documentation for more detail)}

\item{maxAllowedLength}{the maximum length of read that will be aligned (including the sides). regions longer than that will be split and rejoined after alignment}

\item{blastArgs}{advanced: customize the arguments to be passed to blast. by default, blast will be run with with "-max_hsps 1 -max_target_seqs 1 -outfmt 15 -parse_deflines". These are required for the parsing later to work, so please add those in addition to additional argument desired. (user responsibility - no validation by the package)}
}
\value{
A data frame containing the variants with their positions in the new build, based on the following scheme:
\describe{
\item{name}{Name of the variant}
\item{chrom_orig, start_orig, end_orig}{The original position supplied to function.}
\item{chrom, start, end}{Coordinates of alignment in the new build}
}
}
\description{
Lift over coordinates of variants from one genome build to another, by realignment
}
